[TEST_UPDATE=true] Refactoring checklist handling
