+++ modules/thunder_updater/src/Entity/Update.php
@@ -6,7 +6,6 @@ use Drupal\Core\Entity\EntityStorageInterface;
 use Drupal\Core\Field\BaseFieldDefinition;
 use Drupal\Core\Entity\ContentEntityBase;
 use Drupal\Core\Entity\EntityTypeInterface;
-use Drupal\thunder_updater\UpdateInterface;
 
 /**
  * Defines the Contact entity.
