+++ modules/thunder_updater/src/UpdateChecklist.php
@@ -1,0 +1,252 @@
+<?php
+
+namespace Drupal\thunder_updater;
+
+use Drupal\checklistapi\ChecklistapiChecklist;
+use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\Core\Extension\ModuleHandlerInterface;
+use Drupal\Core\Session\AccountInterface;
+use Drupal\thunder_updater\Entity\Update;
+
+/**
+ * Update checklist service.
+ *
+ * @package Drupal\thunder_updater
+ */
+class UpdateChecklist {
+
+  /**
+   * The Checklist API object.
+   *
+   * @var \Drupal\checklistapi\ChecklistapiChecklist
+   */
+  protected $updateChecklist;
+
+  /**
+   * Site configFactory object.
+   *
+   * @var \Drupal\Core\Config\ConfigFactoryInterface
+   */
+  protected $configFactory;
+
+  /**
+   * The account object.
+   *
+   * @var \Drupal\Core\Session\AccountInterface
+   */
+  protected $account;
+
+  /**
+   * Module installer service.
+   *
+   * @var \Drupal\Core\Extension\ModuleHandlerInterface
+   */
+  protected $moduleHandler;
+
+  /**
+   * Update checklist constructor.
+   *
+   * @param \Drupal\Core\Config\ConfigFactoryInterface $configFactory
+   *   Config factory service.
+   * @param \Drupal\Core\Extension\ModuleHandlerInterface $moduleHandler
+   *   Module handler service.
+   * @param \Drupal\Core\Session\AccountInterface $account
+   *   The current user.
+   */
+  public function __construct(ConfigFactoryInterface $configFactory, ModuleHandlerInterface $moduleHandler, AccountInterface $account) {
+    $this->configFactory = $configFactory;
+    $this->moduleHandler = $moduleHandler;
+    $this->account = $account;
+
+    if ($this->moduleHandler->moduleExists('checklistapi')) {
+      $this->updateChecklist = checklistapi_checklist_load('thunder_updater');
+    }
+    else {
+      $this->updateChecklist = FALSE;
+    }
+  }
+
+  /**
+   * Update checklist is available.
+   *
+   * @return bool
+   *   Returns if update checklist is available.
+   */
+  public function isAvailable() {
+    return boolval($this->updateChecklist);
+  }
+
+  /**
+   * Marks a list of updates as successful.
+   *
+   * @param array $names
+   *   Array of update ids.
+   * @param bool $checkListPoints
+   *   Indicates the corresponding checkbox should be checked.
+   */
+  public function markUpdatesSuccessful(array $names, $checkListPoints = TRUE) {
+    foreach ($names as $name) {
+
+      if ($update = Update::load($name)) {
+        $update->setSuccessfulByHook(TRUE)
+          ->save();
+      }
+      else {
+        Update::create([
+          'id' => $name,
+          'successful_by_hook' => TRUE,
+        ])->save();
+      }
+    }
+
+    if ($checkListPoints) {
+      $this->checkListPoints($names);
+    }
+
+  }
+
+  /**
+   * Marks a list of updates as failed.
+   *
+   * @param array $names
+   *   Array of update ids.
+   */
+  public function markUpdatesFailed(array $names) {
+    foreach ($names as $name) {
+
+      if ($update = Update::load($name)) {
+        $update->setSuccessfulByHook(FALSE)
+          ->save();
+      }
+      else {
+        Update::create([
+          'id' => $name,
+          'successful_by_hook' => FALSE,
+        ])->save();
+      }
+    }
+  }
+
+  /**
+   * Marks a list of updates.
+   *
+   * @param bool $status
+   *   Checkboxes enabled or disabled.
+   */
+  public function markAllUpdates($status = TRUE) {
+    if ($this->updateChecklist === FALSE) {
+      return;
+    }
+
+    foreach ($this->updateChecklist->items as $versionItems) {
+      foreach ($versionItems as $key => $item) {
+
+        if (!is_array($item)) {
+          continue;
+        }
+
+        if ($update = Update::load($key)) {
+          $update->setSuccessfulByHook($status)
+            ->save();
+        }
+        else {
+          Update::create([
+            'id' => $key,
+            'successful_by_hook' => $status,
+          ])->save();
+        }
+      }
+    }
+
+    $this->checkAllListPoints($status);
+  }
+
+  /**
+   * Checks an array of bulletpoints on a checklist.
+   *
+   * @param array $names
+   *   Array of the bulletpoints.
+   */
+  protected function checkListPoints(array $names) {
+    if ($this->updateChecklist === FALSE) {
+      return;
+    }
+
+    /** @var \Drupal\Core\Config\Config $thunderUpdaterConfig */
+    $thunderUpdaterConfig = $this->configFactory
+      ->getEditable('checklistapi.progress.thunder_updater');
+
+    $user = $this->account->id();
+    $time = time();
+
+    foreach ($names as $name) {
+      if ($thunderUpdaterConfig && !$thunderUpdaterConfig->get(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . ".#items.$name")) {
+
+        $thunderUpdaterConfig
+          ->set(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . ".#items.$name", [
+            '#completed' => time(),
+            '#uid' => $user,
+          ]);
+
+      }
+    }
+
+    $thunderUpdaterConfig
+      ->set(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . '.#completed_items', count($thunderUpdaterConfig->get(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . ".#items")))
+      ->set(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . '.#changed', $time)
+      ->set(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . '.#changed_by', $user)
+      ->save();
+  }
+
+  /**
+   * Checks all the bulletpoints on a checklist.
+   *
+   * @param bool $status
+   *   Checkboxes enabled or disabled.
+   */
+  protected function checkAllListPoints($status = TRUE) {
+    if ($this->updateChecklist === FALSE) {
+      return;
+    }
+
+    /** @var \Drupal\Core\Config\Config $thunderUpdaterConfig */
+    $thunderUpdaterConfig = $this->configFactory
+      ->getEditable('checklistapi.progress.thunder_updater');
+
+    $user = $this->account->id();
+    $time = time();
+
+    $thunderUpdaterConfig
+      ->set(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . '.#changed', $time)
+      ->set(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . '.#changed_by', $user);
+
+    $exclude = [
+      '#title',
+      '#description',
+      '#weight',
+    ];
+
+    foreach ($this->updateChecklist->items as $versionItems) {
+      foreach ($versionItems as $itemName => $item) {
+        if (!in_array($itemName, $exclude)) {
+          if ($status) {
+            $thunderUpdaterConfig
+              ->set(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . ".#items.$itemName", [
+                '#completed' => $time,
+                '#uid' => $user,
+              ]);
+          }
+          else {
+            $thunderUpdaterConfig
+              ->clear(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . ".#items.$itemName");
+          }
+        }
+      };
+    }
+
+    $thunderUpdaterConfig
+      ->set(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . '.#completed_items', count($thunderUpdaterConfig->get(ChecklistapiChecklist::PROGRESS_CONFIG_KEY . ".#items")))
+      ->save();
+  }
+
+}
